[JSC] JS Lexer should not do double-lookup for dispatch https://bugs.webkit.org/show_bug.cgi?id=296913 rdar://157533153
Reviewed by Keith Miller.

For dispatching, we do not need to lookup the type from
typesOfLatin1Characters. It turned out that this load is really hot, and
we should just directly dispatch the switch based on the coming
character.
The important thing is,

1. We should list up all 0 - 255 cases explicitly to make sure that
   LChar is dispatched somewhere. This leads to efficient code
   generation from the compiler.
2. We should not add the characters outside of 0 - 255 to keep the
   dispatch super efficient. That's why we reuse white space characters
   for non-latin1 identifier dispatching.

* Source/JavaScriptCore/parser/Lexer.cpp:
(JSC::Lexer<T>::lexWithoutClearingLineTerminator):

Canonical link: https://commits.webkit.org/298259@main