FragmentedSharedBuffer should always be immutable. https://bugs.webkit.org/show_bug.cgi?id=296798 rdar://157282731
Reviewed by Gerald Squelart.

Following this change, a FragmentedSharedBuffer is now always immutable
and can never be modified once created.

The SharedBufferBuilder is the object where DataSegmentVector is now kept
and updated.

Upon call to SharedBufferBuilder::take or get, the SharedBuffer or FragmentedSharedBuffer
shared buffer will be lazily created.
Due to how ScriptBuffer uses a SharedBufferBuilder::get() we have to keep this
SharedBuffer alive until the next time a segment is added (or cleared).
This dangerous will be rectified in a follow-up change.

Covered by existing API tests and more added to broaden coverage.

Canonical link: https://commits.webkit.org/298282@main