[Absolute Positioning][Grid] align-self auto should resolve against the parent's align items when statically positioned. https://bugs.webkit.org/show_bug.cgi?id=296903 rdar://problem/157511193
Reviewed by Alan Baradlay.

One of the things we currently look at when determining a box's
eligibility for alignment in its static position rectangle is the value
of its align-self property. If we find that the value is auto then we
return false. This patch builds upon this case by resolving against the
parent's value of align-items in that case and then switching upon the
ItemPosition.

Instead of applying this logic inside of isEligibleForStaticRangeAlignment
directly, we need to add it inside of resolveAlignmentValue. This is
because this function is used by resolveAlignmentShift, which we use to
determine how much we need to move the box to satisfy its alignment, to
determine what the ItemPosition should be. We did not need to modify
this function previously since it was just returning the box's alignment
value which was on m_alignment. Now we resolve against the parent
instead of using ItemPosition::Normal which was occurring before.

When resolving self alignment against the parent we also need to provide
an ItemPosition as what the normal behavior should be. According to
css-align, this would either be "start," or "stretch," depending on the
type of the box with a link to css-position:
https://drafts.csswg.org/css-align-3/#align-abspos

From my reading of css-position, it seems like the only way this can be
Something other than start is if it has non-auto insets (i.e. is not
statically positioned). As a result, we provide ItemPosition::Start as
the normal behavior:
https://drafts.csswg.org/css-position-3/#abspos-layout

resolveAlignmentValue is called indirectly in two different cases: the
first is when we are aligning a box that has fixed insets and fixed
margins into available space and the second is during out new alignment
codepath. Since the first case cannot happened when the box is
statically positioned we can constrain resolveAlignmentValue to only
contain logic that is applicable to cases when the box is statically
positioned and is eligible to be aligned within its static position
rectangle. This is kept in sync with a new member variable that keeps
track of when a box is eligible to take the new codepath. Currently only
compiled in debug builds since it is only used as a way to keep these
two pieces of code in sync.

* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-items-center-large-border-padding.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-items-center.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-items-end-large-border-padding.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-items-end.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-items-flex-end-large-border-padding.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-items-flex-end.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-items-self-end-large-border-padding.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/grid-abspos-staticpos-align-items-self-end.html: Added.
The above tests were added to verify behavior and are the tests we added
in previous patches to test the respective values of align-self but are
not being tested via align-items on the parent.

Canonical link: https://commits.webkit.org/298287@main