EditingStyle: rename forceInline to forceDisplayInline https://bugs.webkit.org/show_bug.cgi?id=297003 rdar://157670437
Reviewed by Sammy Gill.

"forceInline" is not a great name, it is misleading because in the
EditingStyle context "inline" is often used to describe inline style
of elements. It also doesn't express the fact that it is related to the
CSS display property.

* Source/WebCore/editing/EditingStyle.cpp:
(WebCore::EditingStyle::forceDisplayInline):
(WebCore::EditingStyle::forceInline): Deleted.
* Source/WebCore/editing/EditingStyle.h:
* Source/WebCore/editing/markup.cpp:
(WebCore::StyledMarkupAccumulator::appendText):
(WebCore::StyledMarkupAccumulator::appendStartTag):

Canonical link: https://commits.webkit.org/298296@main