[Fuzzilli] Ignore static initializer linker errors when building JavaScriptCore.framework <https://bugs.webkit.org/show_bug.cgi?id=296950> <rdar://157593751>
Reviewed by Elliott Williams and Yusuke Suzuki.

Internal SDKs for iOS 26 and macOS 26 enabled `-no_inits` by default for
targets that are included in the dyld shared cache, which broke the
Fuzzilli build.  This is fixed by adding linker flags to ignore static
initializers and to reassure the linker that the built product won't be
included in a dyld shared cache.

* Configurations/Sanitizers.xcconfig:
(WK_FUZZILLI_OTHER_LDFLAGS_YES):
- Add `-no_warn_inits` and `-not_for_dyld_shared_cache` to fix linking
  of JavaScriptCore.framework with Fuzzilli enabled.

Canonical link: https://commits.webkit.org/298304@main