[Screen Time] Screen Time Website Data may not be cleared properly with multiple Safari profiles https://bugs.webkit.org/show_bug.cgi?id=296967 rdar://157267710
Reviewed by Abrar Rahman Protyasha and Aditya Keerthi.

Website Data is not removed properly because the STWebpageController's
profile identifier and the STWebHistory's profile identifier are
case-sensitive.

Thus, we saw a case where the STWebpageController's profile identifier
was in uppercase but the STWebHistory's profile identifier was in lowercase.
This is because getting the string for a NSUUID is uppercase (STWebpageController)
but getting the string for UUID returns it in lowercase (STWebHistory).

To fix this, we make both profile identifiers the same case for now.
In the future, we should find a way to get the identifier strings in
such a way that does not change the case.

Change the IdentifierString test to match IdentifierStringWithRemoveData.

* Source/WebKit/UIProcess/WebsiteData/Cocoa/ScreenTimeWebsiteDataSupport.mm:
(WebKit::ScreenTimeWebsiteDataSupport::getScreenTimeURLs):
(WebKit::ScreenTimeWebsiteDataSupport::removeScreenTimeData):
(WebKit::ScreenTimeWebsiteDataSupport::removeScreenTimeDataWithInterval):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/ScreenTime.mm:
(TEST(ScreenTime, IdentifierString)):
(TEST(ScreenTime, IdentifierStringWithRemoveData)):

Canonical link: https://commits.webkit.org/298310@main