[iOS] Add development sandboxes https://bugs.webkit.org/show_bug.cgi?id=296999 rdar://157664574
Reviewed by Sihui Liu.

Add development version of sandboxes for iOS. These sandboxes are intended to not block access, but instead
emit reports when a resource that would have been blocked by the non development version is being accessed.
This patch only adds the sandboxes. I will follow up with a patch to make the sandbox changes required to
allow the operations with a sandbox report.

This was first landed in <https://commits.webkit.org/298302@main>. This patch has an additional build fix.

* Source/WebKit/DerivedSources-input.xcfilelist:
* Source/WebKit/DerivedSources-output.xcfilelist:
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.GPU.Development.sb.in: Added.
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.Networking.Development.sb.in: Added.
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.WebContent.Development.sb.in: Added.
* Source/WebKit/Scripts/compile-sandbox.sh:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/298333@main