Restrict location.protocol setter to HTTP(S) schemes https://bugs.webkit.org/show_bug.cgi?id=296972 rdar://157607342
Reviewed by Alex Christensen.

As of 292443@main you can no longer change a URL's scheme from special
to non-special and vice versa. As a side effect of disallowing
non-special to special we would end up navigating to the current
scheme, which would not always work.

So instead let's align with the HTML standard and return early if the
resulting scheme is not an HTTP(S) scheme, which are essentially the
only two schemes for which switching between them makes sense when it
comes to navigation.

As part of this we synchronize WPT
html/browsers/history/the-location-interface and mark the relevant test
that regressed as passing.

Upstream commit: https://github.com/web-platform-tests/wpt/commit/3c34dafeba8c993b3e32ad06450fe4b50499c02e

Canonical link: https://commits.webkit.org/298349@main