[Swift in WebKit] Work towards modularizing JSC private headers (Part 1) https://bugs.webkit.org/show_bug.cgi?id=296551 rdar://156880843
Reviewed by Keith Miller.

This is step 1 towards creating a private module for JavaScriptCore, which in itself
is another step towards facilitating full Swift-Cxx interop in WebKit. Currently, the
WebKit-related modules that exist are:

* WebKit
* WebKit_Private
* WebKit_Internal
* bmalloc
* wtf
* JavaScriptCore

This works towards the introduction of JavaScriptCore_Private; since modules have
stricter requirements than textual includes, headers that belong to a module must be
imported using angle brackets instead of quotation marks. This style of including has
precedent via the previously mentioned modules, and so this also makes things more consistent.

To address this, all of the imports of private headers are changed to use angle
brackets instead of quotation marks. The only other change in this set of changes is to
re-sort the imports, along with these minor related changes:

* JavaScriptCore's CMakeLists.txt and WebKitMacros.cmake is updated to accomodate Linux and Windows builds to work properly with angled bracket imports.

* A few more private headers are made project-level when possible

Canonical link: https://commits.webkit.org/298358@main