Turn `WorkerAsynchronousURLErrorHandlingEnabled` on by default https://bugs.webkit.org/show_bug.cgi?id=297072 <rdar://problem/157769176>
Reviewed by Tim Nguyen.

When we landed the changes in Bug 290745, we had the feature off by default (but on
by default for Safari Technology Preview). After many months of people living on the
feature, it seems like a web compatibility improvement.

Let's turn it on by default for everyone now. The flag will still be present so we can
turn it off if needed, or for specific websites.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:

Canonical link: https://commits.webkit.org/298371@main