RemoteObjectRegistry should decode an invocation with a ReplyBlock if expected. https://bugs.webkit.org/show_bug.cgi?id=290377 rdar://145728621
Reviewed by Alex Christensen.

RemoteObjectRegistry had a logic issue when decoding a RemoteInvocation.

If a CallReplyBlock message was received, this would contain an encoded NSInvocation
representing the method to call and the arguments. It was possible to supply an
alternate selector, with differing arguments, and still have the original ReplyBlock
called with these, leading to a mismatch in argument types.

The RemoteObjectRegistry should ensure the NSInvocation is decoded based on the original
expected ReplyBlock, instead of accepting another selector from the wire.

Originally-landed-as: 289651.466@safari-7621-branch (4a186489a84d). rdar://157795161
Canonical link: https://commits.webkit.org/298380@main