ASAN_TRAP | WebCore::RenderFragmentedFlow::removeRenderBoxFragmentInfo; WebCore::RenderFragmentedFlow::removeFlowChildInfo; WebCore::RenderElement::removeFromRenderFragmentedFlowIncludingDescendants https://bugs.webkit.org/show_bug.cgi?id=288447
Reviewed by Alan Baradlay.

In the test case the video element is added to a fragmented flow, but is not properly removed
when it becomes out of flow due to the popover attribute. The logic in RenderElement::adjustFragmentedFlowStateOnContainingBlockChangeIfNeeded
will not work for this kind of element since the fragmened flow container information is not known anymore, the video having become a top-level element.

To properly remove renderers from their fragmented flow, detect this situation in RenderElement::styleWillChange, and use the fragmented flow access (through locateEnclosingFragmentedFlow())
before the containing block is changed.

* LayoutTests/fast/multicol/video-not-removed-from-fragmented-flow-crash-expected.txt: Added.
* LayoutTests/fast/multicol/video-not-removed-from-fragmented-flow-crash.html: Added.
* Source/WebCore/rendering/RenderElement.cpp:
(WebCore::RenderElement::styleWillChange):

Originally-landed-as: 292955.3@webkit-embargoed (19b66665ae74). rdar://157795098
Canonical link: https://commits.webkit.org/298381@main