Fonts located in application bundles should not be registered in the WebContent process https://bugs.webkit.org/show_bug.cgi?id=297075 rdar://157780044
Reviewed by Sihui Liu.

In <https://commits.webkit.org/298353@main>, we filtered out the fonts local to the app with a string check.
It is better to do this by applying the correct flag to XTCopyPropertiesForAllFontsWithOptions.
The new flag requests that we only get the fonts available globally from the fonts service, not local fonts,
like fonts in apps.

* Source/WebKit/UIProcess/Cocoa/WebProcessPoolCocoa.mm:
(WebKit::WebProcessPool::registerUserInstalledFonts):

Canonical link: https://commits.webkit.org/298392@main