Web Content Crashes when navigating to certain URLs with uBlock origin enabled https://bugs.webkit.org/show_bug.cgi?id=297077 rdar://157688458
Reviewed by Timothy Hatcher and Brian Weinstein.

For certain navigations, we're unnecessarily modifying the request. In testing, it didn't appear
that the actual request was changing, but I suspect something was going wrong here that was causing
some type of corruption.

Since we only have this code to modify the request if a redirect should occur, only apply the
results to the request for this case.

Also, this patch changes the name of redirectedPriorToBlock to redirect and doesn't allow
blockedLoad to be set if a redirect occurred first, and vice versa.

* Source/WebCore/contentextensions/ContentExtensionsBackend.cpp:
(WebCore::ContentExtensions::ContentExtensionsBackend::processContentRuleListsForLoad):
* Source/WebCore/contentextensions/ContentRuleListResults.h:
(WebCore::ContentRuleListResults::shouldBlock const):
* Source/WebCore/loader/PolicyChecker.cpp:
(WebCore::PolicyChecker::checkNavigationPolicy):
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:

Canonical link: https://commits.webkit.org/298398@main