Introduce hit test SPI prototype that works with site isolation https://bugs.webkit.org/show_bug.cgi?id=297079 rdar://157785249
Reviewed by Wenson Hsieh.

This is the initial attempt at making a successor to WKBundleHitTestResultRef and WKWebProcessPlugInHitTestResult,
which can't work with site isolation because they can only return nodes in the same web content process.
This successor returns a _WKNodeInfo that can be passed in to WKWebView.callAsyncJavaScript as a parameter.

In order to get this working, I needed to update some code in EventHandler::hitTestResultAtPoint that assumed
the main frame is in the same process as the frame undergoing the hit test.  This code was introduced in 28403@main,
and moved and optimized in 120022@main.  It seems to be covered by one test,
editing/selection/ios/selection-handle-clamping-in-iframe.html which breaks if it is completely removed.  But for
this hit testing we need to skip it or we will infinitely recurse between main and child frames, so I introduce a new
flag to skip it, HitTestRequest::Type::SkipTransformToRootFrameCoordinates.  I also updated
it to become a little more site isolation friendly by only transforming to the root frame coordinates instead of
assuming the main frame is in the same process, which is not always the case.  This update shouldn't change behavior
with site isolation off because with site isolation off the only root frame is the main frame.

Also, the HitTestRequest::Type used by WKBundleFrameCreateHitTestResult includes AllowChildFrameContent, which
I need to remove in order to get the HTMLFrameOwnerElement to determine if it is local or remote, then continue
in the same process if it is local and in a different process if it is remote.

* Source/WebCore/page/EventHandler.cpp:
(WebCore::EventHandler::hitTestResultAtPoint const):
* Source/WebKit/CMakeLists.txt:
* Source/WebKit/DerivedSources-input.xcfilelist:
* Source/WebKit/DerivedSources.make:
* Source/WebKit/Scripts/webkit/messages.py:
(headers_for_type):
* Source/WebKit/Shared/NodeHitTestResult.h: Added.
* Source/WebKit/Shared/NodeHitTestResult.serialization.in: Added.
* Source/WebKit/UIProcess/API/Cocoa/WKWebView.mm:
(unknownError):
(-[WKWebView createWebArchiveDataWithCompletionHandler:]):
(-[WKWebView _convertPoint:fromFrame:toMainFrameCoordinates:]):
(-[WKWebView _convertRect:fromFrame:toMainFrameCoordinates:]):
(-[WKWebView _hitTestAtPoint:inFrameCoordinateSpace:completionHandler:]):
* Source/WebKit/UIProcess/API/Cocoa/WKWebViewPrivate.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::hitTestAtPoint):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::rootViewToContents):
(WebKit::WebPage::hitTestAtPoint):
* Source/WebKit/WebProcess/WebPage/WebPage.h:
* Source/WebKit/WebProcess/WebPage/WebPage.messages.in:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/SiteIsolation.mm:
(TestWebKitAPI::(SiteIsolation, HitTesting)):

Canonical link: https://commits.webkit.org/298408@main