Simplify WebScriptMessageHandler::Client https://bugs.webkit.org/show_bug.cgi?id=297133 rdar://157869992
Reviewed by Brady Eidson.

Instead of asking the client if it wants a completion handler or not then making one of two very similar calls,
just give the client the completion handler and let it decide when to call it.
This reduces duplicate code and simplifies the future addition of more clients.

* Source/WebKit/UIProcess/API/Cocoa/WKUserContentController.mm:
* Source/WebKit/UIProcess/API/glib/WebKitUserContentManager.cpp:
* Source/WebKit/UIProcess/API/gtk/WebKitRemoteInspectorProtocolHandler.cpp:
* Source/WebKit/UIProcess/Inspector/socket/RemoteInspectorProtocolHandler.cpp:
* Source/WebKit/UIProcess/UserContent/WebScriptMessageHandler.h:
* Source/WebKit/UIProcess/UserContent/WebUserContentControllerProxy.cpp:
(WebKit::WebUserContentControllerProxy::didPostMessage):

Canonical link: https://commits.webkit.org/298412@main