Crash under IDBTransaction::requestPutOrAdd() https://bugs.webkit.org/show_bug.cgi?id=292340 rdar://149937426
Reviewed by Sihui Liu and Per Arne Vollan.

The code was constructing the IDBResultData on the background thread and then
capturing it in the lambda with a WTFMove() before using it on the main thread.
This didn't look thread-safe. We could call `isolatedCopy()` but I have opted
to delay the construction of the IDBResultData until we're on the main thread.

* Source/WebCore/Modules/indexeddb/IDBTransaction.cpp:
(WebCore::IDBTransaction::putOrAddOnServer):

Originally-landed-as: 289651.475@safari-7621-branch (05d9b16c830a). rdar://157794363
Canonical link: https://commits.webkit.org/298424@main