[JSC] [e72bdd394922cc4a] ASAN_ILL | Yarr::YarrPatternConstructor::quantifyAtom; Yarr::Parser::parseTokens; JSC::Yarr::parse https://bugs.webkit.org/show_bug.cgi?id=292305 rdar://149760394
Reviewed by Daniel Liu.

We were inadvertently creating ForwardReference atoms when we should have been creating
BackReference atoms.  This is because we would create a ForwardReference when the referenced
subpatternId was equal to the value of a parenthesis' subpatternId, but that is a valid
BackReference.  The code in question made a >= comparison, but the greater than case was
already handled prior to the code that this change removes.

Added some new regression test cases.

* JSTests/stress/regexp-lookbehind.js:
* Source/JavaScriptCore/yarr/YarrPattern.cpp:
(JSC::Yarr::YarrPatternConstructor::atomBackReference):
(JSC::Yarr::YarrPatternConstructor::atomNamedBackReference):

Originally-landed-as: 289651.472@safari-7621-branch (3665da825356). rdar://157794638
Canonical link: https://commits.webkit.org/298422@main