ASAN_TRAP | WTF::String::utf8; WTR::TestInvocation::dumpResults; WTR::TestInvocation::invoke https://bugs.webkit.org/show_bug.cgi?id=288439
Reviewed by Darin Adler.

Both StringImpl::tryGetUTF8ForCharacters() methods are incorrectly setting a limit
to the size of the vector to be used for the UTF-8 conversion to a fraction
StringImpl::MaxLength, which can cause a valid string not to be converted succesfully
if it's long enough. The purpose of this check should be to prevent overflow
when calculating the size of the vector, so a better approach is to use checked
arithmetic to keep the size within boundaries in a cleaner way.

Some JS tests that rely on overflowing are updated to reflect the change, as the
sizes used in the tests shouldn't really overflow once this is corrected, so use
larger strings instead.

* JSTests/stress/regress-185888.js:
* JSTests/stress/string-16bit-repeat-overflow.js:
* LayoutTests/js/throw-large-string-oom-expected.txt:
* LayoutTests/js/throw-large-string-oom.html:
* Source/WTF/wtf/text/StringImpl.h:
(WTF::StringImpl::tryGetUTF8ForCharacters):

Originally-landed-as: 292955.6@webkit-2025.4-embargoed (aeed21455d86). rdar://157792341
Canonical link: https://commits.webkit.org/298439@main