Fix infinite loop in InsertListCommand::doApply(). https://bugs.webkit.org/show_bug.cgi?id=289376
Reviewed by Ryosuke Niwa.

In https://commits.webkit.org/293666@main, we worked around an infinite
loop in InsertListCommand::doApply() when the "current paragraph" does
not move forward, assuming the old "current paragraph" is still part
of the tree. This latter condition can fail if for example the editor
reformatted the old "current paragraph" and replaced it with a new
element. To be able to check the "current paragraph" does not get
blocked in that case, we instead rely on the position from which we
advanced to that "current paragraph". This allows to addres two cases:

- When we indefinitely listify/unlistify the current paragraph.
- When the indefinitely add new containers around the current paragraph
  (leading to an arbitrary deep tree).

* LayoutTests/fast/editing/insert-list-increasingly-deeper-item-expected.txt: Added.
* LayoutTests/fast/editing/insert-list-increasingly-deeper-item.html: Added.
* LayoutTests/fast/editing/insert-list-keep-listifying-and-unlistifying-indefinitely-expected.txt: Added.
* LayoutTests/fast/editing/insert-list-keep-listifying-and-unlistifying-indefinitely.html: Added.
* LayoutTests/fast/ruby/crash-when-ruby-base-is-collapsed-expected.txt: Rebaselined.
* LayoutTests/fast/text/selection-on-a-detached-tree-expected.txt: Ditto.
* Source/WebCore/editing/InsertListCommand.cpp:
(WebCore::InsertListCommand::doApply): Add an early return ; cherry-pick changes from https://commits.webkit.org/293666@main,
update the selection even when we're bailing out.
* LayoutTests/fast/ruby/crash-when-ruby-base-is-collapsed-expected.txt: Cherry-pick changes from https://commits.webkit.org/293666@main
* LayoutTests/fast/text/selection-on-a-detached-tree-expected.txt: Ditto.

Originally-landed-as: 292955.8@webkit-2025.4-embargoed (af0885753297). rdar://157792000
Canonical link: https://commits.webkit.org/298440@main