[bmalloc] Re-enable use of madvise(MADV_ZERO) for zeroing memory https://bugs.webkit.org/show_bug.cgi?id=297151 rdar://155891567
Reviewed by Yusuke Suzuki.

Previously, this was causing a performance drop because we spent more
time zeroing the very-large wasm memory region; that zeroing is,
however, unnecessary as the memory is already newly allocated and can
thus be zero-allocated directly rather than allocated-and-then-zeroed.
That change has already landed, so we can safely merge this now.

* Source/bmalloc/bmalloc/VMAllocate.h:

Canonical link: https://commits.webkit.org/298469@main