OMG wasm->wasm calls need to restore stack pointer https://bugs.webkit.org/show_bug.cgi?id=293197 rdar://151553598
Reviewed by Keith Miller.

During OMG, we may call a function which tail calls another function, thus
changing the stack pointer. When we return from the tail call, we won't be
in the same stack position, and thus need to restore it. We need to add in
the logic to restore the stack pointer based on the current frame size and
the value of fp.

* JSTests/wasm/stress/omg-tail-call-to-function-with-less-arguments.js: Added.
* Source/JavaScriptCore/wasm/WasmOMGIRGenerator.cpp:
(JSC::Wasm::OMGIRGenerator::addCall):

Originally-landed-as: 289651.530@safari-7621-branch (109e023e12d0). rdar://157791751
Canonical link: https://commits.webkit.org/298478@main