REGRESSION(289946@main): UI process crashes under -[_WKTargetedElementInfo mediaAndLinkURLs] if element has invalid HTTP URL https://bugs.webkit.org/show_bug.cgi?id=297208 rdar://149770223
Reviewed by Aditya Keerthi.

If -targetedElementInfoAt: hit tests to an element with an invalid HTTP
URL (say, an image with a bad src URL), then the UI process will crash
when calling -[_WKTargetedElementInfo mediaAndLinkURLs]. In a post
289946@main world, invalid WTF::URLs convert to `nil` NSURL instances,
and we try to add these nil values to a NSMutableSet object.

This patch works around the crash by checking whether or not we are
adding a nil value first.

API test: ElementTargeting.TargetedElementWithInvalidURLShouldNotCrash.

* Source/WebKit/UIProcess/API/Cocoa/_WKTargetedElementInfo.mm:
(-[_WKTargetedElementInfo mediaAndLinkURLs]):
* Tools/TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* Tools/TestWebKitAPI/Tests/WebKitCocoa/ElementTargetingTests.mm:
(TestWebKitAPI::TEST(ElementTargeting, TargetedElementWithInvalidURLShouldNotCrash)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/element-targeting-13.html: Added.

Canonical link: https://commits.webkit.org/298509@main