[orthogonal td] RenderTableCell::setCellLogicalWidth should set the logical height for orthogonal cells https://bugs.webkit.org/show_bug.cgi?id=297170
Reviewed by Antti Koivisto.

Table sets the final width of the cells by calling RenderTableCell::setCellLogicalWidth.

This is the logical _width_ from the table's point of view which, in case
of orthogonal content is the logical _height_.
This patch ensures that once the table sets this final size, we preserve it throughout
the rest of the layout. This is similar to what we do for non-orthogonal cells by
overriding ::updateLogicalWidth and _ignoring_ all the width computation (essentially an empty function).

* Source/WebCore/rendering/RenderBox.cpp:
(WebCore::RenderBox::computeLogicalHeight const):
* Source/WebCore/rendering/RenderTableCell.cpp:
(WebCore::RenderTableCell::setCellLogicalWidth):
* Source/WebCore/rendering/RenderTableCell.h:
* Source/WebCore/rendering/RenderTableSection.cpp:
(WebCore::RenderTableSection::relayoutCellIfFlexed):

Canonical link: https://commits.webkit.org/298538@main